Fix Invidious URL selection via settings page
* Clean up API URL storage key mis-match. Now properly pulls URL entered
in preferences (originally commented by Sbado in #136) Note that URL in
preferences MUST END WITH A SLASH (may sanitize this later)
* Make default API site invidious.xyz
* Change hardcoded site URLs in search to use URL stored in preferences
(other areas in search and main were already using it)

Tested working in a debug build on my Galaxy S7! ^_^

P.S. I had to release the version lock on cached_network_image to get it
to compile on my system. Hopefully this doesn't cause issues for anyone
else.